rename to config detail
