--- src/main/java/com/tvlk/payment/ruleengine/model/attributes/AttributeTree.java
@@ -1,10 +1,0 @@
-package com.tvlk.payment.ruleengine.model.attributes;
-
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-public class AttributeTree {
-  AttributeNode root;
-}
