+++ src/main/java/com/tvlk/payment/ruleengine/model/configdetail/ConfigDetailTree.java
@@ -1,0 +1,10 @@
+package com.tvlk.payment.ruleengine.model.configdetail;
+
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+public class ConfigDetailTree {
+  ConfigDetail root;
+}
