+++ src/test/resources/config-detail/flight-payment-method-availability-tree.json
