+++ src/test/resources/config-detail/hotel-payment-method-availability.json
